+++ pymongo/common.py
@@ -364,7 +364,7 @@ def validate_read_preference(dummy, value):
 
 
 def validate_read_preference_mode(dummy, value):
-    """Validate read preference mode for a MongoReplicaSetClient.
+    """Validate read preference mode for a MongoClient.
 
     .. versionchanged:: 3.5
        Returns the original ``value`` instead of the validated read preference